+++ src/utils/deltaIdentifier.ts
@@ -1,0 +1,5 @@
+import { ResourceIdentifier } from "@solid/community-server";
+
+export function getDeltaIdentifier(fromIdentifier: ResourceIdentifier): ResourceIdentifier {
+    return { path: fromIdentifier.path + ".vSolid" }
+}
